(baby-project-deployment) #3 Creates a simple deployment action (#5)
I created the pipline prosess for deploying to the server. This only runs when something has been merged to main. The idea in the future is that each PR will create its own instance of the deployment. But that is something we can add in the future.

The PR deployment will be like a RC (Release Candidate). That way we can test all or latest feautres agenst production before we actualy merge it and it gets deployed to production.

But for now we only have the main branch that is essentaly production